# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FXREBATE PROJECT",
        "description": "REACT Progressive Web App / Stand on the shoulders of giants",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/account-types/create": {
            "get": {
                "tags": [
                    "AccountType"
                ],
                "summary": "Get form data for creating account type",
                "operationId": "20a66bba8979a337406542030a04b3ed",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Create form endpoint"
                                        },
                                        "data": {
                                            "properties": {
                                                "brokers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "zones": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "option_categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}